Implement javascript interface
The javascript side is responsible for starting up a nailgun server,
sending requests to it, and shutting it down when no longer required.

The promise disposer pattern is used to manage lifetime of XSLT
executors, allowing users of a nailgun server to be reference counted,
and the server shut down once no longer used.