Restructure XSLTNail command line arguments
It now takes a --system-identifier and <xml-file> option. Instead of
always reading stdin for XML input, it reads <xml-file> if provided,
otherwise SAXON will resolve the system identifier URL. Stdin is read
if <xml-file> is -, or neither --system-identifier or <xml-file> are
specified.

Arguments are now parsed in a more robust way, which is easier to
extend with additional options/arguments.