Refactor AutoCloser implementation
We've now got one primary implementation: DefaultAutoCloser which
delegates to strategy objects to decide when it needs to close its
reference.

This will allow a combination of factors to be used to decide whether
to drop a reference. Specifically, I want to support a timeout to to
keep a reference alive for a period of time after the last reference
was dropped.