Implement a keep-alive timeout for JVM processes
This allows the background nailgun server process to stay running for
some period of time once all execute() calls have finished and all
executor instances are closed. This makes it easier to for independent
sections of a program to share the same nailgun server without
explicitly using the same executor instance.