feat: support stylesheet parameters
This adds a --parameter which accepts values of the form
"{uri}name=value". This option can be repeated to set multiple separate
parameters, or to provide multiple values for the same parameter.

Values are always strings but are exposed to the stylesheet as
xs:untypedAtomic, which results in them being cast to the
parameter's declared type using the standard type conversion rules for
the XSLT stylesheet being executed, e.g:

    https://www.w3.org/TR/xslt-30/#priming-stylesheet
    https://www.w3.org/TR/xpath-30/#id-function-conversion-rules
    https://www.w3.org/TR/xpath-functions-31/#casting
    https://www.w3.org/TR/xpath-functions-31/#casting-from-strings

For string (lexical) representation of the standard data types, see:

    https://www.w3.org/TR/xmlschema-2/#built-in-datatypes