fix: global asyncJVMShutdownError hook wasn't triggered
The hooks.asyncJVMShutdownError hook wasn't being triggered if the
JVMProcess raised in close() when called without a direct caller (e.g
from a timer).

Also, references to JVMProcess instances and their auto-closers were
not being dropped when the process was closed, which could result in
a memory leak if processes were created with different settings each
time. References are now dropped when references to the process are all
closed.