feat(java): Implement automatic server shutdown
The server can now monitor a system process and shut itself down if
that process terminates. The intention is for this to be a safety
mechanism used by processes spawning the server to ensure that server
processes do not remain running if the spawning process is killed or
otherwise fails to shutdown the server.

This feature is used via the --require-running-process=<pid> argument.