fix(java): Fix deadlock in automatic shutdown process
The shutdown hook and the automatic shutdown manager were deadlocking
as the shutdown hook ended up blocked on the automatic shutdown
completing as a side-effect of waiting for the server shutdown future
to complete. That couldn't complete because System.exit was called
in the automatic shutdown, and that blocks forever, but never actually
exits the JVM because it was invoked from a shutdown hook.

Now the automatic shutdown and shutdown futures use async for their
steps, so actions are not executed on the threads of callers (such as
the shutdown hook).