test: use jest new/modern fake timers in all tests
This fixes the XSLTExecutor jvmKeepAliveTimeout test failing under jest
27 (the current version).

The jest legacy timers are being removed in 28, so we need to stop
using them.

The XSLTExecutor jvmKeepAliveTimeout test did not work using the new
fake timers. It's not a great fit for the fake timers use case, as it
depends on real time elapsing in order to work (waiting for an external
process to start). So it was rather fiddly to get it to work. I ended
up with a hybrid approach where real time ticks while starting the
server and performing a transform, then time can be advanced
artificially in-between executions.

I also had to move the XSLExecutor test using fake timers into a
separate module as I couldn't find a way to stop the fake and real
timers interacting between tests, despite using the method documented
by jest to re-enable real timers after each test.